requirements.txt
discord.py>=2.0.0
aiohttp>=3.8.0
pandas>=1.5.0
numpy>=1.23.0
python-dotenv>=0.19.0
